pyproject.toml
[build-system]
requires = ["setuptools >= 42.0.0", "wheel"]
build-backend = "setuptools.build_meta:__legacy__"

[project]
name = "ansys-tools-omniverse"
version = "1.0.0"
description = "A kit extension to send data from Ansys products into Omniverse"
readme = "README.rst"
requires-python = "==3.10"
license = {file = "LICENSE"}
authors = [{name = "ANSYS, Inc.", email = "pyansys.core@ansys.com"}]
maintainers = [{name = "ANSYS, Inc.", email = "pyansys.core@ansys.com"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
tests = [
    "ansys-pyensight-core~=0.8.11",
    "omniverse_kit~=106.0.0",
    "pytest==8.3.2",
    "pytest-mock==3.10.0",
]

[tool.pip]
extra-index-url = "https://pypi.nvidia.com/"

[tool.black]
line-length = 100
src_paths = ["doc", "exts"]

[tool.codespell]
ignore-words-list = "ro, te, pres"
quiet-level = 3
skip = "*.pyc,*.xml,*.gif,*.png,*.jpg,*.js,*.html,*.svg"

[tool.isort]
profile = "black"
skip_gitignore = true
force_sort_within_sections = true
line_length = 100
default_section = "THIRDPARTY"
filter_files = true
src_paths = ["ansys", "doc", "tests", "exts"]

[tool.bandit]
level = "HIGH"
number = "3"
recursive = true
exclude = ["venv/*", "tests/*"]

[pydocstyle]
convention = "numpy"
